Added Try/Except structure to call for results
Added a Try/Except/Else If structure in the call to retrieve the experimental results, so if there is a minor network glitch the program will not halt, but can try again to retrieve the results. At time of upload the maximum number of cycles achieved without interruption was > 8900 repeated calls to the simulator. That run only halted when the IBM QX was briefly taken down for maintenance.