updated 5 and 16 qubit versions using giskit
These updated versions both use the full qiskit library for authentication and quantum circuit execution. In addition, the display engine for the output has been generalized to be identical for either; the display list is set to match one of the two globals, and the display engine figures out what to do from there.