New version compatible with v0.9 and v0.12
This version is compatible with the authentication techniques used in both Qiskit 0.9.0 (using qiskit-ibmq-provider v0.2) and Qiskit 0.12.0 (using qiskit-ibmq-provider v0.3).   
It checks the version of the provider module and then creates the object Q with the correct authentication methods, either IBMQ (v0.2) or provider0 (v0.3).
The methods and properties used in the rest of the code are identical in both versions of Qiskit, so once authentication is done the program can proceed regardless of which is installed.